pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["tuppey_tony <miro_19@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
python-dotenv = {extras = ["cli"], version = "1.0.1"}
uvicorn = "0.27.0"
gunicorn = "21.2.0"
alembic = "1.13.1"
psycopg2 = "2.9.9"
asyncpg = "0.29.0"
pydantic = {version = "<2.5.0", extras = ["email"]}
pydantic-settings = "2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "2.0.25"}
fastapi = "0.110.0"
async-fastapi-jwt-auth = "0.6.4"
werkzeug = "3.0.1"
orjson = "3.9.15"
redis-om = "0.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "7.0.0"
flake8-blind-except = "0.2.1"
flake8-broken-line = "1.0.0"
flake8-bugbear = "24.1.17"
flake8-builtins = "2.2.0"
flake8-class-attributes-order = "0.1.3"
flake8-cognitive-complexity = "0.1.0"
flake8-commas = "2.1.0"
flake8-comprehensions = "3.14.0"
flake8-debugger = "4.1.2"
flake8-eradicate = "1.5.0"
flake8-functions = "0.0.8"
flake8-isort = "6.1.1"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-pytest = "1.4"
flake8-pytest-style = "1.7.2"
flake8-quotes = "3.3.2"
flake8-string-format = "0.3.0"
flake8-variables-names = "0.0.6"
isort = "5.13.2"
flake8-docstrings = "1.7.0"
autopep8 = "2.0.4"
pre-commit = "3.6.0"
bandit = "1.7.7"
mypy = "1.8.0"
sqlalchemy = {extras = ["mypy"], version = "2.0.25"}
autoflake = "2.3.0"
flake8-unused-arguments = "0.0.13"


[tool.poetry.group.test.dependencies]
pytest = "8.0.1"
pytest-asyncio = "0.23.5"
httpx = "0.27.0"
asgi-lifespan = "2.1.0"
pytest-cov = "4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
